+++ DustyPig.TMDB/Models/TvSeries/Latest.cs
@@ -52,7 +52,7 @@ public class Latest
     public string Overview { get; set; }
 
     [JsonPropertyName("popularity")]
-    public int Popularity { get; set; }
+    public float Popularity { get; set; }
 
     [JsonPropertyName("seasons")]
     public List<Seasons2> Seasons { get; set; } = [];
